+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - MongoDB: Do not treat the string NULL as the NULL value
 - MongoDB: Show the primary key column when altering indexes (regression from 5.4.0)
 - Redis: SQL command in the redis-cli syntax, print executed commands
+- Redis: Print binary keys and values in the redis-cli syntax
 - SimpleDB: Use default server
 - CSS: Improve contrast of query time and dark mode
 - Editor: Display all images recognized by PHP, set their size, load them lazily